requirements.txt
streamlit
google-generativeai
beautifulsoup4
lxml
cloudscraper
tenacity
brotli
curl_cffi
